requirements.txt
strands-agents 
strands-agents-tools
flask
flask_cors
python-dotenv
boto3
requests